2 phase commit

+++ central_server.py
@@ -10,10 +10,12 @@ import sys
 import socket
 from pymongo import MongoClient
 from datetime import datetime
+import logging
 
 if len(sys.argv) < 3 : 
     print "ERROR : Enter the port and type of server (0 for master, 1 for backup)...exiting"
     exit()
+
 port = sys.argv[1]
 IS_MASTER = False
 if sys.argv[2] == "0" :
@@ -21,51 +23,67 @@ if sys.argv[2] == "0" :
 _ONE_DAY_IN_SECONDS = 60 * 60 * 24
 SELF_IP=[l for l in ([ip for ip in socket.gethostbyname_ex(socket.gethostname())[2] if not ip.startswith("127.")][:1], [[(s.connect(('8.8.8.8', 53)), s.getsockname()[0], s.close()) for s in [socket.socket(socket.AF_INET, socket.SOCK_DGRAM)]][0][1]]) if l][0][0]
 
-def two_phase_commit(response):
-    if response.ack=="Agreed":
-        print "master writes in db"
-        log.write(str(datetime.now())+" commit ")
-        channel = grpc.insecure_channel(allotedServerBackup)
-        stub = pr_pb2_grpc.PublishTopicStub(channel)
-        response_phase2 = stub.commit_phase_two(pr_pb2.empty())
-        print response_phase2.ack
+def two_phase_init(request):
+    channel = grpc.insecure_channel(backupCentralServer)
+    stub = pr_pb2_grpc.PublishTopicStub(channel)
+    response = stub.commit_request(request)
+    if response.ack=="AGREED":
+        logging.info("%s:%s:COMMIT",str(datetime.now()),request.filename)
+        response1 = stub.commit_phase_two(request)
+        return response1.ack
+    else :
+        return "ERROR"
+        
 
 class CentralServer(pr_pb2_grpc.PublishTopicServicer):
     def commit_request(self,request,context):
-        # dct = json.load(open("dataBackup/centralDictionary"+ str(port),"r"))
+        if request.filename == "twoLevelDict":        
+            if request.action == "remove" :
+                if request.level == "3" : 
+                    twoLevelDict.delete_one({"topic":request.data_1,"subscriber":request.data_3})
+                    logging.info("%s:%s:REMOVE 3 %s %s %s",str(datetime.now()),request.filename,request.data_1,"NIL",request.data_3)
+                    
+                elif request.level == "2" :
+                    twoLevelDict.delete_many({"topic":request.data_1,"publisher":request.data_2})
+                    logging.info("%s:%s:REMOVE 2 %s %s %s",str(datetime.now()),request.filename,request.data_1,request.data_2,"NIL")
+                    
+            elif request.action == "insert":
+                if request.level == "3" :
+                   twoLevelDict.insert_one({"topic":request.data_1,"publisher":request.data_2,"subscriber":request.data_3}) 
+                   logging.info("%s:%s:INSERT 3 %s %s %s",str(datetime.now()),request.filename,request.data_1,request.data_2,request.data_3)
+                   
+        elif request.filename == "frontends" :
+            if request.action == "remove" :
+                if request.level == "2" : 
+                    if request.data_1 == "ip" :
+                        frontends.delete_one({"type":request.data_1,request.data_1:request.data_2})
+                        logging.info("%s:%s:REMOVE 2 %s %s %s",str(datetime.now()),request.filename,request.data_1,request.data_2,request.data_3)
+                    else :
+                        frontends.delete_one({"type":request.data_1,request.data_1:int(request.data_2)})
+                        logging.info("%s:%s:REMOVE 2 %s %s %s",str(datetime.now()),request.filename,request.data_1,request.data_2,request.data_3)
 
-        if request.level == "1":
-            if request.action == "add":
-                pass
-            else:
-                pass
-        elif request.level== "2":
-            if request.action == "add":
-                pass
-            else:
-                pass
-        else:
-            if request.action == "add":
-                print "add ho rha hai.."
-            else:
-                print "level 3 remove ho rha"
-        # json.dump(dct,open(request.filename,"w"))
-        # a=json.dumps(request)
-        log.write(str(datetime.now())+" dataBackup/centralDictionary "+ str(port))
-        print SELF_IP
-        return pr_pb2.acknowledge(ack="Agreed")
+            elif request.action == "insert" :
+                if request.level == "2" : 
+                    if request.data_1 == "ip" :
+                        frontends.insert_one({"type":request.data_1,request.data_1:request.data_2})
+                        logging.info("%s:%s:REMOVE 2 %s %s %s",str(datetime.now()),request.filename,request.data_1,request.data_2,request.data_3)
+                    else :
+                        frontends.insert_one({"type":request.data_1,request.data_1:int(request.data_2)})
+                        logging.info("%s:%s:REMOVE 2 %s %s %s",str(datetime.now()),request.filename,request.data_1,request.data_2,request.data_3)
+        
+        return pr_pb2.acknowledge(ack="AGREED")
 
     def commit_phase_two(self,request,context):
-        log.write(str(datetime.now())+" commit")
-        return pr_pb2.acknowledge(ack="completed")
+        logging.info("%s:%s:COMMIT",str(datetime.now()),request.filename)
+        return pr_pb2.acknowledge(ack="COMPLETE")
 
     def unsubscribeRequestCentral(self, request, context):
-        twoLevelDict.delete_one({"topic":request.topic,"subscriber":request.client_ip})
         if IS_MASTER:
-            channel = grpc.insecure_channel(allotedServerBackup)
-            stub = pr_pb2_grpc.PublishTopicStub(channel)
-            response = stub.commit_request(pr_pb2.commit_req_data(action="remove",level="3",data_1 = request.topic,data_2="LOL", data_3 = request.client_ip, filename = "db"))
-            response1 = two_phase_commit(response)
+            response = two_phase_init(pr_pb2.commit_req_data(action="remove",level="3",data_1 = request.topic,data_2="", data_3 = request.client_ip, filename = "twoLevelDict",function_name="unsubscribeRequestCentral"))
+            if response == "COMPLETE" :
+                twoLevelDict.delete_one({"topic":request.topic,"subscriber":request.client_ip})
+        else :
+            twoLevelDict.delete_one({"topic":request.topic,"subscriber":request.client_ip})
         return pr_pb2.acknowledge(ack="temporary acknowledge from central server")
 
     def deReplicaRequest(self, request, context):
@@ -80,7 +98,13 @@ class CentralServer(pr_pb2_grpc.PublishTopicServicer):
         if len(dct.keys()) == 1:
             return pr_pb2.acknowledge(ack="ERROR")
         extraSubscribers = dct[request.client_ip]
-        twoLevelDict.delete_many({"topic":request.topic,"publisher":request.client_ip})
+        if IS_MASTER:
+            response = two_phase_init(pr_pb2.commit_req_data(action="remove",level="2",data_1 = request.topic,data_2=request.client_ip, data_3 = "", filename = "twoLevelDict",function_name="deReplicaRequest"))
+            if response == "COMPLETE" :
+                twoLevelDict.delete_many({"topic":request.topic,"publisher":request.client_ip})
+        else :
+            twoLevelDict.delete_many({"topic":request.topic,"publisher":request.client_ip})
+        # twoLevelDict.delete_many({"topic":request.topic,"publisher":request.client_ip})
         del dct[request.client_ip]
         for subscriber in extraSubscribers :
             allotedServer = ""
@@ -92,7 +116,13 @@ class CentralServer(pr_pb2_grpc.PublishTopicServicer):
                     l = tempCursor.count()
                     tempIp = ip
 
-            twoLevelDict.insert_one({"topic":request.topic,"publisher":tempIp,"subscriber":subscriber})
+            if IS_MASTER:
+                response = two_phase_init(pr_pb2.commit_req_data(action="insert",level="3",data_1 = request.topic,data_2=tempIp, data_3 = subscriber, filename = "twoLevelDict",function_name="deReplicaRequest"))
+                if response == "COMPLETE" :
+                    twoLevelDict.insert_one({"topic":request.topic,"publisher":tempIp,"subscriber":subscriber})
+                else :
+                    twoLevelDict.insert_one({"topic":request.topic,"publisher":tempIp,"subscriber":subscriber})
+            # twoLevelDict.insert_one({"topic":request.topic,"publisher":tempIp,"subscriber":subscriber})
         return pr_pb2.acknowledge(ack="DONE") 
 
     def querryTopics(self, request, context):
@@ -106,10 +136,28 @@ class CentralServer(pr_pb2_grpc.PublishTopicServicer):
         if twoLevelDict.find({"topic":request.topic,"publisher":request.client_ip}).count() > 0 :
             return pr_pb2.acknowledge(ack="Requesting front end server already a replica for "+request.topic)
         if twoLevelDict.find({"topic":request.topic,"subscriber":request.client_ip}).count() > 0 :
-            twoLevelDict.delete_one({"topic":request.topic,"subscriber":request.client_ip})
+            if IS_MASTER:
+                response = two_phase_init(pr_pb2.commit_req_data(action="remove",level="3",data_1 = request.topic,data_2="", data_3 = request.client_ip, filename = "twoLevelDict",function_name="replicaRequest"))
+                if response == "COMPLETE" :
+                    twoLevelDict.delete_one({"topic":request.topic,"subscriber":request.client_ip})
+                else :
+                    twoLevelDict.delete_one({"topic":request.topic,"subscriber":request.client_ip})
+            # twoLevelDict.delete_one({"topic":request.topic,"subscriber":request.client_ip})
 
-        twoLevelDict.insert_one({"topic":request.topic,"publisher":request.client_ip,"subscriber":"NULL"})
-        twoLevelDict.insert_one({"topic":request.topic,"publisher":request.client_ip,"subscriber":request.client_ip})
+        if IS_MASTER:
+            response = two_phase_init(pr_pb2.commit_req_data(action="insert",level="3",data_1 = request.topic,data_2=request.client_ip, data_3 = "NULL", filename = "twoLevelDict",function_name="replicaRequest"))
+            if response == "COMPLETE" :
+                twoLevelDict.insert_one({"topic":request.topic,"publisher":request.client_ip,"subscriber":"NULL"})
+            else :
+                twoLevelDict.insert_one({"topic":request.topic,"publisher":request.client_ip,"subscriber":"NULL"})
+        # twoLevelDict.insert_one({"topic":request.topic,"publisher":request.client_ip,"subscriber":"NULL"})
+        if IS_MASTER:
+            response = two_phase_init(pr_pb2.commit_req_data(action="insert",level="3",data_1 = request.topic,data_2=request.client_ip, data_3 = request.client_ip, filename = "twoLevelDict",function_name="replicaRequest"))
+            if response == "COMPLETE" :
+                twoLevelDict.insert_one({"topic":request.topic,"publisher":request.client_ip,"subscriber":request.client_ip})
+            else :
+                twoLevelDict.insert_one({"topic":request.topic,"publisher":request.client_ip,"subscriber":request.client_ip})
+        # twoLevelDict.insert_one({"topic":request.topic,"publisher":request.client_ip,"subscriber":request.client_ip})
         channel = grpc.insecure_channel(allotedServer)
         stub = pr_pb2_grpc.PublishTopicStub(channel)
         response = stub.sendBackupRequestReplica(pr_pb2.topicSubscribe(topic=request.topic, client_ip=request.client_ip))
@@ -132,7 +180,13 @@ class CentralServer(pr_pb2_grpc.PublishTopicServicer):
                         l=twoLevelDict.find({"topic":request.topic,"publisher":ip}).count()
                         tempIp = ip
                 allotedServer = tempIp
-            twoLevelDict.insert_one({"topic":request.topic,"publisher":allotedServer,"subscriber":request.client_ip})
+            if IS_MASTER:
+                response = two_phase_init(pr_pb2.commit_req_data(action="insert",level="3",data_1 = request.topic,data_2=allotedServer, data_3 = request.client_ip, filename = "twoLevelDict",function_name="subscribeRequestCentral"))
+                if response == "COMPLETE" :
+                    twoLevelDict.insert_one({"topic":request.topic,"publisher":allotedServer,"subscriber":request.client_ip})
+                else :
+                    twoLevelDict.insert_one({"topic":request.topic,"publisher":allotedServer,"subscriber":request.client_ip})
+            # twoLevelDict.insert_one({"topic":request.topic,"publisher":allotedServer,"subscriber":request.client_ip})
 
         else :
             document = twoLevelDict.find_one({"topic":request.topic,"subscriber":request.client_ip})
@@ -165,7 +219,13 @@ class CentralServer(pr_pb2_grpc.PublishTopicServicer):
             for document in cursor :
                 lst.append(document["ip"])
             ip = random.choice(lst)
-            twoLevelDict.insert_one({"topic":request.topic,"publisher":ip,"subscriber":"NULL"})
+            if IS_MASTER:
+                response = two_phase_init(pr_pb2.commit_req_data(action="insert",level="3",data_1 = request.topic,data_2=ip, data_3 = "NULL", filename = "twoLevelDict",function_name="giveIps"))
+                if response == "COMPLETE" :
+                    twoLevelDict.insert_one({"topic":request.topic,"publisher":ip,"subscriber":"NULL"})
+                else :
+                    twoLevelDict.insert_one({"topic":request.topic,"publisher":ip,"subscriber":"NULL"})
+            # twoLevelDict.insert_one({"topic":request.topic,"publisher":ip,"subscriber":"NULL"})
             yield pr_pb2.ips(ip=ip)
 
     def getFrontIp(self, request, context) :
@@ -181,16 +241,54 @@ class CentralServer(pr_pb2_grpc.PublishTopicServicer):
             ipList.append(document["ip"])
         m = ipList[index]
         if index == len(ipList) - 1 :
-            frontends.replace_one({"type":"index","index":index},{"type":"index","index":0})
+            if IS_MASTER:
+                response = two_phase_init(pr_pb2.commit_req_data(action="remove",level="2",data_1 = "index",data_2=str(index), data_3 = "", filename = "frontends",function_name="getFrontIp"))
+                if response == "COMPLETE" :
+                    frontends.delete_one({"type":"index","index":index})
+                else :
+                    frontends.delete_one({"type":"index","index":index})
+            # frontends.delete_one({"type":"index","index":index})
+            if IS_MASTER:
+                response = two_phase_init(pr_pb2.commit_req_data(action="insert",level="2",data_1 = "index",data_2="0", data_3 = "", filename = "frontends",function_name="getFrontIp"))
+                if response == "COMPLETE" :
+                    frontends.insert_one({"type":"index","index":0})
+                else :
+                    frontends.insert_one({"type":"index","index":0})
+            # frontends.insert_one({"type":"index","index":0})
         else :
-            frontends.replace_one({"type":"index","index":index},{"type":"index","index":index+1})
+            if IS_MASTER:
+                response = two_phase_init(pr_pb2.commit_req_data(action="remove",level="2",data_1 = "index",data_2=str(index), data_3 = "", filename = "frontends",function_name="getFrontIp"))
+                if response == "COMPLETE" :
+                    frontends.delete_one({"type":"index","index":index})
+                else :
+                    frontends.delete_one({"type":"index","index":index})
+            # frontends.delete_one({"type":"index","index":index})
+            if IS_MASTER:
+                response = two_phase_init(pr_pb2.commit_req_data(action="insert",level="2",data_1 = "index",data_2=str(index+1), data_3 = "", filename = "frontends",function_name="getFrontIp"))
+                if response == "COMPLETE" :
+                    frontends.insert_one({"type":"index","index":index+1})
+                else :
+                    frontends.insert_one({"type":"index","index":index+1})
+            # frontends.insert_one({"type":"index","index":index+1})
         return pr_pb2.ips(ip=m)
 
     def registerIp(self, request, context) :
         cursor = frontends.find({"type":"index"})
         if cursor.count() == 0:
-            frontends.insert_one({"type":"index","index":0})
-        frontends.insert_one({"type":"ip","ip":request.ip})
+            if IS_MASTER:
+                response = two_phase_init(pr_pb2.commit_req_data(action="insert",level="2",data_1 = "index",data_2="0", data_3 = "", filename = "frontends",function_name="getFrontIp"))
+                if response == "COMPLETE" :
+                    frontends.insert_one({"type":"index","index":0})
+                else :
+                    frontends.insert_one({"type":"index","index":0})
+            # frontends.insert_one({"type":"index","index":0})
+        if IS_MASTER:
+            response = two_phase_init(pr_pb2.commit_req_data(action="insert",level="2",data_1 = "ip",data_2=request.ip, data_3 = "", filename = "frontends",function_name="getFrontIp"))
+            if response == "COMPLETE" :
+                frontends.insert_one({"type":"ip","ip":request.ip})
+            else :
+                frontends.insert_one({"type":"ip","ip":request.ip})
+        # frontends.insert_one({"type":"ip","ip":request.ip})
         return pr_pb2.acknowledge(ack="Ip added...")
 
 
@@ -207,15 +305,10 @@ def serve():
 
 
 if __name__ == '__main__':
-    # selfIpDct = {}
-    # if IS_MASTER :
-    #     selfIpDct["centralServer"] = str(SELF_IP)+":"+port
-    # else :
-    #     selfIpDct["centralServerBackup"] = str(SELF_IP)+":"+port
-    log=open('log.txt'+port,'w')
-    # json.dump(selfIpDct,open("options","w"))
+    logging.basicConfig(filename=port+'.log',format='%(message)s',filemode='w',level=logging.DEBUG)
+    print SELF_IP
     selfIpDct = json.load(open("options","r"))
-    allotedServerBackup = selfIpDct["centralServerBackup"]
+    backupCentralServer = selfIpDct["centralServerBackup"]
     mongoClient = MongoClient("localhost", 27017)
     mongoClient.drop_database('CentralServer'+port)
     db = mongoClient['CentralServer'+port]
